Adapt to moving of files into KafkaW subdirectory
