+++ src/CommandListener.cxx
@@ -1,5 +1,4 @@
 #include "CommandListener.h"
-#include "KafkaW.h"
 #include "helper.h"
 #include "logger.h"
