+++ src/logger.cxx
@@ -1,5 +1,5 @@
 #include "logger.h"
-#include "KafkaW.h"
+#include "KafkaW/KafkaW.h"
 #include <atomic>
 #include <cstdarg>
 #include <cstdio>
