+++ src/send-command.cxx
@@ -1,4 +1,4 @@
-#include "KafkaW.h"
+#include "KafkaW/KafkaW.h"
 #include "logger.h"
 #include <cstdio>
 #include <cstdlib>
