+++ src/tests/roundtrip.cxx
@@ -1,5 +1,5 @@
 #include "roundtrip.h"
-#include "../KafkaW.h"
+#include "../KafkaW/KafkaW.h"
 #include "../Master.h"
 #include "../Msg.h"
 #include "../helper.h"
